A bit of reorganization and generalization of database endpoints.
